Add files via upload
In this project, I'm going to explore the phenomenon using age distribution data to see if we can reproduce a difference in average age at death purely from the changing rates of left-handedness over time, refuting the claim of early death for left-handers. This notebook uses pandas and Bayesian statistics to analyse the probability of being a certain age at death given that you are reported as left-handed or right-handed.